Refactor from_sgrid_conventions to model
Split out data manipulation and grid creation from the construction of fields
